gocos: rename library to golog_adapter to make name available for app
