+++ src/CMakeLists.txt
@@ -5,7 +5,8 @@ add_subdirectory(utilities)
 add_subdirectory(search)
 add_subdirectory(visualization)
 add_subdirectory(app)
+
 if(TACOS_GOCOS)
-  add_subdirectory(gocos)
+  add_subdirectory(golog_adapter)
   add_subdirectory(golog_app)
 endif()
