+++ src/golog_adapter/golog_symbols.cpp
@@ -6,7 +6,7 @@
  *  SPDX-License-Identifier: LGPL-3.0-or-later
  ****************************************************************************/
 
-#include "gocos/golog_symbols.h"
+#include "golog_adapter/golog_symbols.h"
 
 #include <regex>
 #include <stdexcept>
