+++ src/golog_adapter/include/golog_adapter/golog_symbols.h
