+++ test/benchmark_golog_robot.cpp
@@ -6,8 +6,8 @@
  *  SPDX-License-Identifier: LGPL-3.0-or-later
  ****************************************************************************/
 #include "benchmark_golog.h"
-#include "gocos/golog_adapter.h"
-#include "gocos/golog_program.h"
+#include "golog_adapter/golog_adapter.h"
+#include "golog_adapter/golog_program.h"
 #include "golog_robot.h"
 #include "heuristics_generator.h"
 #include "mtl/MTLFormula.h"
