+++ test/test_golog_words.cpp
@@ -6,8 +6,8 @@
  *  SPDX-License-Identifier: LGPL-3.0-or-later
  ****************************************************************************/
 
-#include "gocos/golog_adapter.h"
-#include "gocos/golog_program.h"
+#include "golog_adapter/golog_adapter.h"
+#include "golog_adapter/golog_program.h"
 #include "mtl/MTLFormula.h"
 #include "mtl_ata_translation/translator.h"
